backend/user-service/requirement.txt
fastapi==0.100.0
uvicorn
motor
pydantic[email]==2.0.1
python-jose[cryptography]
passlib
bcrypt